requirements.txt
docling>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
numpy>=1.21.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0